requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.13.0
asgiref==3.11.1
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
Django==6.0.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.40.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.14.0
packaging==26.1
pluggy==1.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.13.0
pydantic_core==2.46.0
PyJWT==2.12.1
pytest==8.3.5
pytest-django==4.10.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
